+++ app/services/create_journey_step.rb
@@ -3,8 +3,8 @@ class CreateJourneyStep
 
   class UnexpectedContentfulStepType < StandardError; end
 
-  ALLOWED_CONTENTFUL_MODELS = %w[question].freeze
-  ALLOWED_CONTENTFUL_QUESTION_TYPES = ["radios", "short_text", "long_text"].freeze
+  ALLOWED_CONTENTFUL_MODELS = %w[question staticContent].freeze
+  ALLOWED_CONTENTFUL_ENTRY_TYPES = %w[radios short_text long_text paragraphs].freeze
 
   attr_accessor :journey, :contentful_entry
   def initialize(journey:, contentful_entry:)
@@ -26,6 +26,8 @@ class CreateJourneyStep
     step = Step.create(
       title: title,
       help_text: help_text,
+      body: body,
+      contentful_model: content_model,
       contentful_type: step_type,
       options: options,
       raw: raw,
@@ -56,7 +58,7 @@ class CreateJourneyStep
   end
 
   def expected_contentful_step_type?
-    ALLOWED_CONTENTFUL_QUESTION_TYPES.include?(step_type)
+    ALLOWED_CONTENTFUL_ENTRY_TYPES.include?(step_type)
   end
 
   def unexpected_contentful_step_type?
@@ -70,6 +72,11 @@ class CreateJourneyStep
   def help_text
     return nil unless contentful_entry.respond_to?(:help_text)
     contentful_entry.help_text
+  end
+
+  def body
+    return nil unless contentful_entry.respond_to?(:body)
+    contentful_entry.body
   end
 
   def options
@@ -100,7 +107,7 @@ class CreateJourneyStep
       content_model: content_model,
       step_type: step_type,
       allowed_content_models: ALLOWED_CONTENTFUL_MODELS.join(", "),
-      allowed_step_types: ALLOWED_CONTENTFUL_QUESTION_TYPES.join(", ")
+      allowed_step_types: ALLOWED_CONTENTFUL_ENTRY_TYPES.join(", ")
     )
   end
 end