+++ db/migrate/20201202105320_add_body_to_steps.rb
@@ -1,0 +1,5 @@
+class AddBodyToSteps < ActiveRecord::Migration[6.0]
+  def change
+    add_column :steps, :body, :text
+  end
+end