+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_11_30_155247) do
+ActiveRecord::Schema.define(version: 2020_12_02_150359) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -56,6 +56,8 @@ ActiveRecord::Schema.define(version: 2020_11_30_155247) do
     t.binary "raw", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.text "body"
+    t.string "contentful_model"
     t.index ["journey_id"], name: "index_steps_on_journey_id"
   end