+++ spec/factories/step.rb
@@ -8,20 +8,30 @@ FactoryBot.define do
 
     trait :radio do
       options { ["Red", "Green", "Blue"] }
+      contentful_model { "question" }
       contentful_type { "radios" }
       association :radio_answer
     end
 
     trait :short_text do
       options { nil }
+      contentful_model { "question" }
       contentful_type { "short_text" }
       association :short_text_answer
     end
 
     trait :long_text do
       options { nil }
+      contentful_model { "question" }
       contentful_type { "long_text" }
       association :long_text_answer
+    end
+
+    trait :static_content do
+      options { nil }
+      contentful_model { "staticContent" }
+      contentful_type { "paragraphs" }
+      association :paragraph_content
     end
   end
 end